api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
redis[hiredis]==5.2.1
arq==0.26.1
httpx==0.28.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
